+++ src/tfjs/tf-node-wasm.ts
@@ -1,0 +1,2 @@
+export * from '@tensorflow/tfjs';
+export * from '@tensorflow/tfjs-backend-wasm';