+++ test/test-node.js
@@ -50,36 +50,51 @@ async function testInstance(human) {
     log.error('failed: load models');
   }
 
-  const warmup = await human.warmup();
+  let warmup;
+  try {
+    warmup = await human.warmup();
+  } catch (err) {
+    log.error('error warmup');
+  }
   if (warmup) {
     log.state('passed: warmup:', config.warmup);
-    log.data(' result: face:', warmup.face.length, 'body:', warmup.body.length, 'hand:', warmup.hand.length, 'gesture:', warmup.gesture.length, 'object:', warmup.object.length);
-    log.data(' result: performance:', 'load:', warmup.performance.load, 'total:', warmup.performance.total);
+    log.data(' result: face:', warmup.face?.length, 'body:', warmup.body?.length, 'hand:', warmup.hand?.length, 'gesture:', warmup.gesture?.length, 'object:', warmup.object?.length);
+    log.data(' result: performance:', 'load:', warmup.performance?.load, 'total:', warmup.performance?.total);
   } else {
     log.error('failed: warmup');
   }
   const random = tf.randomNormal([1, 1024, 1024, 3]);
-  const detect = await human.detect(random);
+  let detect;
+  try {
+    detect = await human.detect(random);
+  } catch (err) {
+    log.error('error: detect', err);
+  }
   tf.dispose(random);
   if (detect) {
     log.state('passed: detect:', 'random');
-    log.data(' result: face:', detect.face.length, 'body:', detect.body.length, 'hand:', detect.hand.length, 'gesture:', detect.gesture.length, 'object:', detect.object.length);
-    log.data(' result: performance:', 'load:', detect.performance.load, 'total:', detect.performance.total);
+    log.data(' result: face:', detect.face?.length, 'body:', detect.body?.length, 'hand:', detect.hand?.length, 'gesture:', detect.gesture?.length, 'object:', detect.object?.length);
+    log.data(' result: performance:', 'load:', detect?.performance.load, 'total:', detect.performance?.total);
   } else {
     log.error('failed: detect');
   }
 }
 
 async function test() {
-  log.info('testing instance#1');
-  config.warmup = 'face';
+  log.info('testing instance#1 - none');
+  config.warmup = 'none';
   const human1 = new Human(config);
   await testInstance(human1);
 
-  log.info('testing instance#2');
-  config.warmup = 'body';
+  log.info('testing instance#2 - face');
+  config.warmup = 'face';
   const human2 = new Human(config);
   await testInstance(human2);
+
+  log.info('testing instance#3 - body');
+  config.warmup = 'body';
+  const human3 = new Human(config);
+  await testInstance(human3);
 }
 
 test();