+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # @vladmandic/human  
 
-Version: **1.5.0**  
+Version: **1.5.1**  
 Description: **Human: AI-powered 3D Face Detection & Rotation Tracking, Face Description & Recognition, Body Pose Tracking, 3D Hand & Finger Tracking, Iris Analysis, Age & Gender & Emotion Prediction, Gesture Recognition**  
 
 Author: **Vladimir Mandic <mandic00@live.com>**  
@@ -9,8 +9,10 @@ Repository: **<git+https://github.com/vladmandic/human.git>**
 
 ## Changelog
 
-### **HEAD -> main** 2021/04/13 mandic00@live.com
+### **1.5.1** 2021/04/13 mandic00@live.com
 
+- fix for safari imagebitmap
+- refactored human.config and human.draw
 
 ### **1.4.3** 2021/04/12 mandic00@live.com